requirements.txt
aiosmtpd
bootstrap-flask
dnspython
flask
flask-sqlalchemy
flask-wtf
flask-login
pymysql
python-ldap
